Tests/LibWeb/Text/expected/HTML/Window-find-inert.txt

window.find("inert") initial result: true
window.find("inert") second call: false
